did the full implementation

+++ game/game_test.go
@@ -6,14 +6,14 @@ import (
 )
 
 func TestNew(t *testing.T) {
-	g := New(map[string]uint{"t1": 0, "t2": 0, "t3": 3}, []byte("salt"), nil)
+	g := New(map[string]uint64{"t1": 0, "t2": 0, "t3": 3}, []byte("salt"), nil)
 	if g.winnings == nil {
 		t.Errorf("Did not initialize winnings map")
 	}
 }
 
 func TestGame_Play(t *testing.T) {
-	g := New(map[string]uint{"t1": 0, "t2": 0, "t3": 3}, []byte("salt"), crypto.NewRng())
+	g := New(map[string]uint64{"t1": 0, "t2": 0, "t3": 3}, []byte("salt"), crypto.NewRng())
 	ok, _ := g.Play("addr", "i'm a message")
 	if ok {
 		t.Error("Should not have been able to play with unknown address")
